requirements.txt
apache-beam==2.53.0
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
crcmod==1.7
dill==0.3.1.1
dnspython==2.3.0
docopt==0.6.2
fasteners==0.18
Flask==2.3.2
grpcio==1.54.2
hdfs==2.7.0
httplib2==0.22.0
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.24.3
objsize==0.6.1
orjson==3.9.7
proto-plus==1.22.2
protobuf==4.25.2
py4j==0.10.9.7
pyarrow==11.0.0
pydot==1.4.2
pymongo==4.3.3
pyparsing==3.0.9
pyspark==3.4.0
python-dateutil==2.8.2
pytz==2023.3
regex==2023.6.3
requests==2.31.0
six==1.12.0
typing_extensions==4.6.3
urllib3==2.0.3
Werkzeug==2.3.6
zipp==3.15.0
zstandard==0.21.0
Flask-SQLAlchemy==3.0.3
PyMySQL==1.0.3
alembic==1.11.1
flasgger==0.9.7.1
google-cloud-pubsub==2.17.1
SQLAlchemy==2.0.25
pydantic==2.6.2
pydantic_core==2.16.3
matplotlib==3.8.4
matplotlib-inline==0.1.6
scipy==1.13.0
Flask-CQLAlchemy==2.0.0
cqlengine==0.21.0
python-dotenv==1.0.1